Refactor highlight.py out of existence
Related to #542
