+++ docs/index.rst
@@ -17,11 +17,6 @@ File
     
 .. automodule:: file
 
-Highlight
-=========
-
-.. automodule:: highlight
-
 MainText
 ========
