+++ src/guiguts/word_frequency.py
@@ -8,7 +8,6 @@ from typing import Any, Callable
 import regex as re
 
 from guiguts.file import PAGE_SEPARATOR_REGEX
-from guiguts.highlight import spotlight_range, remove_spotlights
 from guiguts.maintext import maintext
 from guiguts.mainwindow import ScrolledReadOnlyText
 from guiguts.misc_tools import tool_save
@@ -683,10 +682,10 @@ class WordFrequencyDialog(ToplevelDialog):
                     if match_str[1 : match.count + 1] == newline_word:
                         match.rowcol.col += 1
             maintext().set_insert_index(match.rowcol, focus=False)
-            remove_spotlights()
+            maintext().remove_spotlights()
             start_index = match.rowcol.index()
             end_index = maintext().index(start_index + f"+{match.count}c")
-            spotlight_range(IndexRange(start_index, end_index))
+            maintext().spotlight_range(IndexRange(start_index, end_index))
             self.previous_word = word
 
     def search_word(self, event: tk.Event) -> str:
